PLATO-000: move language selector component into features folder
