+++ src/components/features/language-selector/index.ts
