+++ src/components/templates/header/header.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from 'next-i18next';
 import Link from 'next/link';
 
 import BlogLogo from '@icons/blog-logo.svg';
-import { LanguageSelector } from '@src/components/language-selector';
+import { LanguageSelector } from '@src/components/features/language-selector';
 import { Container } from '@src/components/shared/container';
 
 export const Header = () => {
